+++ command/encoding/json_test.go
@@ -7,11 +7,12 @@ import (
 )
 
 func Test_JSONNoEscaping(t *testing.T) {
+	enc := Encoder{}
 	m := map[string]string{
 		"a": "b",
 		"c": "d ->> e",
 	}
-	b, err := JSONMarshal(m)
+	b, err := enc.JSONMarshal(m)
 	if err != nil {
 		t.Fatalf("failed to marshal simple map: %s", err.Error())
 	}
@@ -25,13 +26,14 @@ func Test_MarshalExecuteResult(t *testing.T) {
 	var b []byte
 	var err error
 	var r *command.ExecuteResult
+	enc := Encoder{}
 
 	r = &command.ExecuteResult{
 		LastInsertId: 1,
 		RowsAffected: 2,
 		Time:         1234,
 	}
-	b, err = JSONMarshal(r)
+	b, err = enc.JSONMarshal(r)
 	if err != nil {
 		t.Fatalf("failed to marshal ExecuteResult: %s", err.Error())
 	}
@@ -46,7 +48,7 @@ func Test_MarshalExecuteResult(t *testing.T) {
 		Time:         6789,
 	}
 
-	b, err = JSONMarshal(r)
+	b, err = enc.JSONMarshal(r)
 	if err != nil {
 		t.Fatalf("failed to marshal ExecuteResult: %s", err.Error())
 	}
@@ -54,7 +56,7 @@ func Test_MarshalExecuteResult(t *testing.T) {
 		t.Fatalf("failed to marshal ExecuteResult: exp %s, got %s", exp, got)
 	}
 
-	b, err = JSONMarshalIndent(r, "", "    ")
+	b, err = enc.JSONMarshalIndent(r, "", "    ")
 	if err != nil {
 		t.Fatalf("failed to marshal ExecuteResult: %s", err.Error())
 	}
@@ -74,6 +76,7 @@ func Test_MarshalExecuteResult(t *testing.T) {
 func Test_MarshalExecuteResults(t *testing.T) {
 	var b []byte
 	var err error
+	enc := Encoder{}
 
 	r1 := &command.ExecuteResult{
 		LastInsertId: 1,
@@ -85,7 +88,7 @@ func Test_MarshalExecuteResults(t *testing.T) {
 		RowsAffected: 4,
 		Time:         5678,
 	}
-	b, err = JSONMarshal([]*command.ExecuteResult{r1, r2})
+	b, err = enc.JSONMarshal([]*command.ExecuteResult{r1, r2})
 	if err != nil {
 		t.Fatalf("failed to marshal ExecuteResults: %s", err.Error())
 	}
@@ -99,6 +102,7 @@ func Test_MarshalQueryRows(t *testing.T) {
 	var b []byte
 	var err error
 	var r *command.QueryRows
+	enc := Encoder{}
 
 	r = &command.QueryRows{
 		Columns: []string{"c1", "c2", "c3"},
@@ -126,7 +130,7 @@ func Test_MarshalQueryRows(t *testing.T) {
 		{Parameters: values},
 	}
 
-	b, err = JSONMarshal(r)
+	b, err = enc.JSONMarshal(r)
 	if err != nil {
 		t.Fatalf("failed to marshal QueryRows: %s", err.Error())
 	}
@@ -134,7 +138,7 @@ func Test_MarshalQueryRows(t *testing.T) {
 		t.Fatalf("failed to marshal QueryRows: exp %s, got %s", exp, got)
 	}
 
-	b, err = JSONMarshalIndent(r, "", "    ")
+	b, err = enc.JSONMarshalIndent(r, "", "    ")
 	if err != nil {
 		t.Fatalf("failed to marshal QueryRows: %s", err.Error())
 	}
@@ -164,11 +168,14 @@ func Test_MarshalQueryRows(t *testing.T) {
 	}
 }
 
-// Test_MarshalQueryRowses tests JSON marshaling of a slice of QueryRows
-func Test_MarshalQueryRowses(t *testing.T) {
+// Test_MarshalQueryAssociativeRows tests JSON marshaling of a QueryRows
+func Test_MarshalQueryAssociativeRows(t *testing.T) {
 	var b []byte
 	var err error
 	var r *command.QueryRows
+	enc := Encoder{
+		Associative: true,
+	}
 
 	r = &command.QueryRows{
 		Columns: []string{"c1", "c2", "c3"},
@@ -196,11 +203,116 @@ func Test_MarshalQueryRowses(t *testing.T) {
 		{Parameters: values},
 	}
 
-	b, err = JSONMarshal([]*command.QueryRows{r, r})
+	b, err = enc.JSONMarshal(r)
+	if err != nil {
+		t.Fatalf("failed to marshal QueryRows: %s", err.Error())
+	}
+	if exp, got := `{"rows":[{"c1":123,"c2":678,"c3":"fiona"}],"time":6789}`, string(b); exp != got {
+		t.Fatalf("failed to marshal QueryRows: exp %s, got %s", exp, got)
+	}
+
+	b, err = enc.JSONMarshalIndent(r, "", "    ")
+	if err != nil {
+		t.Fatalf("failed to marshal QueryRows: %s", err.Error())
+	}
+	exp := `{
+    "rows": [
+        {
+            "c1": 123,
+            "c2": 678,
+            "c3": "fiona"
+        }
+    ],
+    "time": 6789
+}`
+	got := string(b)
+	if exp != got {
+		t.Fatalf("failed to pretty marshal QueryRows: exp: %s, got: %s", exp, got)
+	}
+}
+
+// Test_MarshalQueryRowses tests JSON marshaling of a slice of QueryRows
+func Test_MarshalQueryRowses(t *testing.T) {
+	var b []byte
+	var err error
+	var r *command.QueryRows
+	enc := Encoder{}
+
+	r = &command.QueryRows{
+		Columns: []string{"c1", "c2", "c3"},
+		Types:   []string{"int", "float", "string"},
+		Time:    6789,
+	}
+	values := make([]*command.Parameter, len(r.Columns))
+	values[0] = &command.Parameter{
+		Value: &command.Parameter_I{
+			I: 123,
+		},
+	}
+	values[1] = &command.Parameter{
+		Value: &command.Parameter_D{
+			D: 678.0,
+		},
+	}
+	values[2] = &command.Parameter{
+		Value: &command.Parameter_S{
+			S: "fiona",
+		},
+	}
+
+	r.Values = []*command.Values{
+		{Parameters: values},
+	}
+
+	b, err = enc.JSONMarshal([]*command.QueryRows{r, r})
 	if err != nil {
 		t.Fatalf("failed to marshal QueryRows: %s", err.Error())
 	}
 	if exp, got := `[{"columns":["c1","c2","c3"],"types":["int","float","string"],"values":[[123,678,"fiona"]],"time":6789},{"columns":["c1","c2","c3"],"types":["int","float","string"],"values":[[123,678,"fiona"]],"time":6789}]`, string(b); exp != got {
+		t.Fatalf("failed to marshal QueryRows: exp %s, got %s", exp, got)
+	}
+}
+
+// Test_MarshalQueryRowses tests JSON marshaling of a slice of QueryRows
+func Test_MarshalQueryAssociativeRowses(t *testing.T) {
+	var b []byte
+	var err error
+	var r *command.QueryRows
+	enc := Encoder{
+		Associative: true,
+	}
+
+	r = &command.QueryRows{
+		Columns: []string{"c1", "c2", "c3"},
+		Types:   []string{"int", "float", "string"},
+		Time:    6789,
+	}
+	values := make([]*command.Parameter, len(r.Columns))
+	values[0] = &command.Parameter{
+		Value: &command.Parameter_I{
+			I: 123,
+		},
+	}
+	values[1] = &command.Parameter{
+		Value: &command.Parameter_D{
+			D: 678.0,
+		},
+	}
+	values[2] = &command.Parameter{
+		Value: &command.Parameter_S{
+			S: "fiona",
+		},
+	}
+
+	r.Values = []*command.Values{
+		{Parameters: values},
+	}
+
+	b, err = enc.JSONMarshal([]*command.QueryRows{r, r})
+	if err != nil {
+		t.Fatalf("failed to marshal QueryRows: %s", err.Error())
+	}
+	if exp, got := `[{"rows":[{"c1":123,"c2":678,"c3":"fiona"}],"time":6789},{"rows":[{"c1":123,"c2":678,"c3":"fiona"}],"time":6789}]`, string(b); exp != got {
 		t.Fatalf("failed to marshal QueryRows: exp %s, got %s", exp, got)
 	}
 }