+++ db/db_test.go
@@ -1890,7 +1890,8 @@ func mustQuery(db *DB, stmt string) {
 }
 
 func asJSON(v interface{}) string {
-	b, err := encoding.JSONMarshal(v)
+	enc := encoding.Encoder{}
+	b, err := enc.JSONMarshal(v)
 	if err != nil {
 		panic(fmt.Sprintf("failed to JSON marshal value: %s", err.Error()))
 	}