+++ store/store_test.go
@@ -2182,7 +2182,8 @@ func waitForLeaderID(s *Store, timeout time.Duration) (string, error) {
 }
 
 func asJSON(v interface{}) string {
-	b, err := encoding.JSONMarshal(v)
+	enc := encoding.Encoder{}
+	b, err := enc.JSONMarshal(v)
 	if err != nil {
 		panic(fmt.Sprintf("failed to JSON marshal value: %s", err.Error()))
 	}